django/requirements.txt
asgiref==3.2.7
awscli==1.16.272
bson==0.5.8
certifi==2020.4.5.1
dataclasses==0.7
Django==2.2.12
django-filter==2.2.0
djangorestframework==3.11.0
djongo==1.3.2
docopt==0.6.2
joblib==0.14.1
numpy==1.18.3
Pillow==7.1.1
pymongo==3.10.1
python-dateutil==2.8.1
pythonanywhere==0.8.1
pytz==2019.3
rsa==3.4.2
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.2.4
gunicorn
drf_yasg
django-heroku